# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "安然画室 · 课程预约小程序 API",
    "version": "1.0.0",
    "description": "线下课程介绍+预约微信小程序后端接口文档。包含用户端公开接口、用户端需登录接口、管理后台接口三部分。\n\n**认证方式**: JWT Bearer Token（Header: `Authorization: Bearer <token>`）\n\n**管理后台**: 用户名密码登录，使用 nickmame 作为用户名，默认密码为环境变量 `ADMIN_PASSWORD`（默认 `admin123`）\n\n**响应格式**: 统一为 `{ code: 0, message: \"ok\", data: {} }`\n\n**开发测试账号**: openid=dev_test_openid，无需真实微信code即可登录"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.anran-studio.com/api",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，登录后获取。用户端使用微信code登录，管理端使用用户名密码登录。"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误描述"
          },
          "data": {
            "type": "null",
            "description": "始终为 null"
          }
        }
      },
...
      "CourseSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string",
            "description": "课程标题"
          },
          "cover": {
            "type": "string",
            "description": "封面图URL"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "OrderSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "course_id": {
            "type": "integer"
          },
          "course_type_id": {
            "type": "integer"
          },
          "schedule_id": {
            "type": "integer"
          },
          "student_id": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "cancelled"
            ],
            "description": "订单状态"
          },
          "audit_remark": {
            "type": "string",
            "nullable": true,
            "description": "审核备注"
          },
          "pay_status": {
            "type": "string",
            "enum": [
              "unpaid",
              "paid",
              "refunded"
            ],
            "description": "支付状态（预留）"
          },
          "transaction_id": {
            "type": "string",
            "nullable": true
          },
          "course": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "title": {
                "type": "string"
              },
              "cover": {
                "type": "string"
              }
            }
          },
          "courseType": {
            "$ref": "#/components/schemas/CourseType"
          },
          "schedule": {
            "$ref": "#/components/schemas/CourseSchedule"
          },
          "student": {
            "$ref": "#/components/schemas/Student"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "description": "无管理员权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": 403,
              "message": "无管理员权限",
              "data": null
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/courses": {
      "get": {
        "tags": [
          "课程"
        ],
        "summary": "课程列表（分页+筛选+搜索）",
        "description": "获取已发布的课程列表，支持按分类筛选和关键词搜索。",
        "security": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1,
              "example": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页条数（最大100）",
            "schema": {
              "type": "integer",
              "default": 10,
              "example": 10
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "按分类ID筛选",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "关键词搜索（匹配课程标题）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "课程状态筛选",
            "schema": {
              "type": "string",
              "enum": [
                "published",
                "draft",
                "archived"
              ],
              "default": "published"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页课程列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CourseSummary"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "nullable": true
                    }
                  }
                },
                "example": {
                  "code": 0,
                  "message": "ok",
                  "data": {
                    "list": [
                      {
                        "id": 1,
                        "title": "Scratch 少儿编程启蒙班",
                        "cover": "",
                        "status": "published",
                        "category": {
                          "id": 1,
                          "name": "少儿编程"
                        },
                        "created_at": "2026-05-01T00:00:00.000Z"
                      }
                    ],
                    "total": 1,
                    "page": 1,
                    "pageSize": 10,
                    "totalPages": 1
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/orders": {
      "post": {
        "tags": [
          "预约订单"
        ],
        "summary": "提交预约",
        "description": "用户提交课程预约。检查时段名额（行级锁防超卖）和同一用户重复提交。成功后booked_count+1。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "course_id",
                  "course_type_id",
                  "schedule_id",
                  "student_id"
                ],
                "properties": {
                  "course_id": {
                    "type": "integer",
                    "description": "课程ID"
                  },
                  "course_type_id": {
                    "type": "integer",
                    "description": "班型ID"
                  },
                  "schedule_id": {
                    "type": "integer",
                    "description": "时段ID"
                  },
                  "student_id": {
                    "type": "integer",
                    "description": "学员ID"
                  }
                }
              },
              "example": {
                "course_id": 1,
                "course_type_id": 1,
                "schedule_id": 1,
                "student_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "预约提交成功，状态为pending",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "预约已提交，请等待管理员审核"
                    },
                    "data": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/OrderSummary"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数缺失/名额已满/重复提交",
            "content": {
              "application/json": {
                "examples": {
                  "missingParams": {
                    "value": {
                      "code": 400,
                      "message": "缺少必填参数",
                      "data": null
                    }
                  },
                  "duplicate": {
                    "value": {
                      "code": 400,
                      "message": "该时段您已有预约，请勿重复提交",
                      "data": null
                    }
                  },
                  "cancelled": {
                    "value": {
                      "code": 400,
                      "message": "该时段已取消",
                      "data": null
                    }
                  },
                  "full": {
                    "value": {
                      "code": 400,
                      "message": "名额已满",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "tags": [
          "预约订单"
        ],
        "summary": "我的预约列表",
        "description": "获取当前用户的预约订单列表，支持按状态筛选和分页。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "按状态筛选（不传返回全部）",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页订单列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderSummary"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/students": {
      "get": {
        "tags": [
          "学员管理"
        ],
        "summary": "我的学员列表",
        "description": "获取当前用户添加的所有学员档案。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "学员列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Student"
                      },
                      "nullable": true
                    }
                  }
                },
                "example": {
                  "code": 0,
                  "message": "ok",
                  "data": [
                    {
                      "id": 1,
                      "user_id": 1,
                      "name": "朵朵",
                      "age": 5,
                      "phone": "13800138000",
                      "notes": "",
                      "created_at": "2026-05-01T00:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "学员管理"
        ],
        "summary": "添加学员",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "学员姓名",
                    "example": "朵朵"
                  },
                  "age": {
                    "type": "integer",
                    "description": "年龄",
                    "default": 0,
                    "example": 5
                  },
                  "phone": {
                    "type": "string",
                    "description": "联系电话",
                    "default": "",
                    "example": "13800138000"
                  },
                  "notes": {
                    "type": "string",
                    "description": "备注信息",
                    "default": "",
                    "example": "对美术特别感兴趣"
                  }
                }
              },
              "example": {
                "name": "朵朵",
                "age": 5,
                "phone": "13800138000",
                "notes": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "添加成功"
                    },
                    "data": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Student"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "姓名不能为空",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "学员姓名不能为空",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/admin/login": {
      "post": {
        "tags": [
          "管理端-认证"
        ],
        "summary": "管理后台登录",
        "description": "使用用户名（nickname）和密码登录管理后台。默认管理员：username=admin，password=admin123（环境变量ADMIN_PASSWORD控制）。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "管理员用户名（即用户的nickname字段，role=admin）",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "description": "管理员密码",
                    "example": "admin123"
                  }
                }
              },
              "example": {
                "username": "admin",
                "password": "admin123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "description": "JWT Token"
                        },
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      },
                      "nullable": true
                    }
                  }
                },
                "example": {
                  "code": 0,
                  "message": "ok",
                  "data": {
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "user": {
                      "id": 1,
                      "openid": "admin_test_openid",
                      "nickname": "admin",
                      "avatar": "",
                      "phone": "",
                      "role": "admin",
                      "created_at": "2026-05-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "缺少用户名密码",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "请输入用户名和密码",
                  "data": null
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误",
            "content": {
              "application/json": {
                "example": {
                  "code": 401,
                  "message": "用户名或密码错误",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/admin/dashboard": {
      "get": {
        "tags": [
          "管理端-仪表盘"
        ],
        "summary": "仪表盘核心数据",
        "description": "获取管理后台仪表盘统计数据：今日订单、待审核数、学员总数、课程总数、近30天预约趋势。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "仪表盘数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/DashboardData"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                },
                "example": {
                  "code": 0,
                  "message": "ok",
                  "data": {
                    "todayOrders": 5,
                    "pendingCount": 3,
                    "totalStudents": 120,
                    "totalCourses": 8,
                    "trend": [
                      {
                        "date": "2026-04-01",
                        "count": 10
                      },
                      {
                        "date": "2026-04-02",
                        "count": 8
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/admin/courses": {
      "get": {
        "tags": [
          "管理端-课程"
        ],
        "summary": "课程列表（含草稿和归档）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "筛选状态",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "archived"
              ]
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "关键词搜索课程标题",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页课程列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CourseSummary"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "管理端-课程"
        ],
        "summary": "创建课程",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "category_id"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "课程名称",
                    "example": "Scratch 少儿编程启蒙班"
                  },
                  "category_id": {
                    "type": "integer",
                    "description": "分类ID",
                    "example": 1
                  },
                  "cover": {
                    "type": "string",
                    "description": "封面图URL",
                    "example": ""
                  },
                  "intro": {
                    "type": "string",
                    "description": "课程介绍（HTML富文本）",
                    "example": "<h3>课程介绍</h3><p>适合6-10岁...</p>"
                  },
                  "teacher_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "description": "讲师ID数组",
                    "example": [
                      1
                    ]
                  },
                  "location_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "description": "上课地点ID数组",
                    "example": [
                      1,
                      2
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published",
                      "archived"
                    ],
                    "description": "课程状态",
                    "default": "draft",
                    "example": "draft"
                  }
                }
              },
              "example": {
                "title": "Scratch 少儿编程启蒙班",
                "category_id": 1,
                "cover": "",
                "intro": "<h3>课程介绍</h3><p>适合6-10岁零基础儿童...</p>",
                "teacher_ids": [
                  1
                ],
                "location_ids": [
                  1,
                  2
                ],
                "status": "draft"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "创建成功"
                    },
                    "data": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/CourseDetail"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "课程名称和分类为必填",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "课程名称和分类为必填项",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/admin/orders": {
      "get": {
        "tags": [
          "管理端-订单"
        ],
        "summary": "预约订单列表",
        "description": "获取所有用户的预约订单，含用户、课程、班型、时段、学员信息。支持按状态筛选和分页。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "按状态筛选",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "cancelled"
              ]
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索（目前仅用于前端过滤，后端未实现搜索逻辑）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页订单列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderAdmin"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  }
}